% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aconr.R
\docType{package}
\name{aconr-package}
\alias{aconr}
\alias{aconr-package}
\title{Active Config Data Structure}
\description{
This package provides the \code{\link{acon}} class that implements a hierachical datastructure
containing key-value pairs at each level of the hierarchy, with
each value defined as R expressions that can reference keys from both the current
and upper levels, with the expression recomputed at each access to the key using
current values of other keys.
}
\details{
The main entry point is the constructor of the \code{\link{acon}}
class.
}
\author{
\strong{Maintainer}: Andrey Tovchigrechko \email{andreyto@gmail.com}

}
