% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acon.R
\name{as.list.acon}
\alias{as.list.acon}
\title{Create a hierarchical named list with the same structure
as the provided Active Config \code{\link{acon}} object,
and with the values computed from the active config.}
\usage{
\method{as.list}{acon}(x, ...)
}
\arguments{
\item{x}{\code{\link{acon}} object}

\item{...}{other parameters are ignored; they are here to maintain signature
of the \code{\link[base]{as.list}} generic}
}
\description{
\strong{Note}: Active bindings are evaluated and replaced with the returned
values. Values that are functions will currently keep their original environment,
even when their environment is the \code{acon} object that is being converted here.
}
\examples{
z_acon = acon(v=4,w=acon(x=v*2))
z_acon
z_list = as.list(z_acon)
z_list
z_list$w$x == 8
}
