% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acon.R
\name{print.acon}
\alias{print.acon}
\title{Print \code{\link{acon}} object}
\usage{
\method{print}{acon}(x, max.level = 100, ...)
}
\arguments{
\item{x}{\code{\link{acon}} object}

\item{max.level}{maximum hierarchy level to print}

\item{...}{other parameters passed to the print function}
}
\description{
Print \code{\link{acon}} object
}

